assessmentproject/requirement.txt
django==3.0.5
djangorestframework==3.11.0
social-auth-app-django==3.1.0
djangorestframework-simplejwt==4.4.0
psycopg2==2.7.3.2
gunicorn==19.7.1
dj-database-url==0.4.2
whitenoise==3.2
